% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Read Input}
\usage{
read_input(ps_inputfile)
}
\arguments{
\item{ps_inputfile}{name of the input file}
}
\value{
tibble with input
}
\description{
Read the input from a file
}
